Organization and small bug fix
Organization - All files were moved and unecessary files were deleted,
as they are created at run time.
bug fix - Instead of saving just the file name of the comic It was
saving the original location.
All the code regarding the feed reader was deleted as the amount of
comics in it's history was to small.

KNOWN BUGS - For some reason the first comic to be download isn't being
download, in both first time and update modes.

TODO - I'm still researching how to make a GUI for the reader part and
making this software more user friendly.
Divide the code in more files and maybe make it OO and/or more
encapsulated.