BUG FIXES and improvements
The bug where it wouldn't download the first required comic is fixed, as
well as some improvements in the download code.
It isn't necessary to explicity say that it is the first time you use
it, when updating, if it finds an empty database it start from the first
comic.