JRuby and Gui
There was a minor modification made for the code tu run on JRuby. In
the future there will be added instructions to change back to the usual
Ruby interpreter on the documentation.

The change to JRuby was made because Java has a well documented graphic
library.

There will be no addition of code that runs solely on Jruby that
download and manage that people can use it freely.